requirements.txt
alembic==0.9.3
amqp==2.2.2
aniso8601==1.2.1
APScheduler==3.4.0
asn1crypto==0.23.0
astroid==1.5.3
autopep8==1.3.3
billiard==3.5.0.3
blinker==1.4
celery==4.1.0
certifi==2017.11.5
cffi==1.11.2
chardet==3.0.4
click==6.7
coverage==4.0.3
cryptography==2.1.4
decorator==4.2.1
dominate==2.3.4
ecdsa==0.13
Flask==0.12.2
Flask-Bootstrap==3.3.7.1
Flask-JWT==0.3.2
Flask-Login==0.4.1
Flask-Migrate==2.0.4
Flask-Script==2.0.5
Flask-SQLAlchemy==2.2
Flask-SSE==0.2.1
Flask-SSLify==0.1.5
Flask-Testing==0.6.2
Flask-WTF==0.14.2
future==0.16.0
gevent==1.2.2
gitdb==0.6.4
GitPython==0.3.2rc1
greenlet==0.4.12
gunicorn==19.7.1
idna==2.6
isort==4.2.15
itsdangerous==0.24
Jinja2==2.9.6
pluggy==0.5.2
psycopg2==2.7.1
pycrypto==2.6.1
PyJWT==1.5.2
pylint==1.7.4
pytest==3.1.3
pytest-cov==2.5.1
python-coveralls==2.9.1
python-dateutil==2.6.0
python-dotenv==0.6.4
python-editor==1.0.3
python-jose==1.3.2
pytz==2017.2
PyYAML==3.12
raven==6.5.0
redis==2.10.6
requests==2.18.4
six==1.10.0
smmap==0.9.0
SQLAlchemy==1.1.14
validators==0.12.1
vine==1.1.4
visitor==0.1.3
voluptuous==0.10.5
walrus==0.4.1
webargs==1.7.0
Werkzeug==0.12.2
wrapt==1.10.11
WTForms==2.2.1
Flask-Babel==0.12.2
